// packages/react-line-flex/src/types.ts
export type FlexSize = 'xxs' | 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl' | '3xl' | '4xl' | '5xl';
export type FlexSpacing = 'none' | 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl';
export type FlexMargin = FlexSpacing;
export type FlexGravity = 'top' | 'bottom' | 'center';
export type FlexAlign = 'start' | 'end' | 'center';
export type FlexDecoration = 'none' | 'underline' | 'line-through';
export type FlexWeight = 'regular' | 'bold';
export type FlexStyle = 'normal' | 'italic';
export type FlexPosition = 'relative' | 'absolute';
export type FlexAspectRatio = '1:1' | '1.51:1' | '1.91:1' | '4:3' | '16:9' | '20:13' | '2:1' | '3:1' | '3:4' | '9:16' | '1:2' | '1:3';
export type FlexAspectMode = 'cover' | 'fit';
// ...
export type FlexAction = FlexURIAction | FlexMessageAction | FlexPostbackAction;

export interface FlexURIAction {
  type: 'uri';
  label?: string;
  uri: string;
  altUri?: {
    desktop?: string;
  };
}

export interface FlexMessageAction {
  type: 'message';
  label?: string;
  text: string;
}

export interface FlexPostbackAction {
  type: 'postback';
  label?: string;
  data: string;
  displayText?: string;
}
// ...
export interface FlexImage {
  type: 'image';
  url: string;
  flex?: number;
  margin?: FlexMargin | string;
  position?: FlexPosition;
  offsetTop?: string;
  offsetBottom?: string;
  offsetStart?: string;
  offsetEnd?: string;
  align?: FlexAlign;
  gravity?: FlexGravity;
  size?: FlexSize | string;
  aspectRatio?: FlexAspectRatio;
  aspectMode?: FlexAspectMode;
  backgroundColor?: string;
  action?: FlexAction;
}